src/agents/requirements.txt
azure-identity>=1.20.0
python-dotenv>=1.0.1
semantic-kernel==1.22.0
dapr>=1.14.0
fastapi>=0.115.11
uvicorn>=0.34.0
dapr-ext-fastapi>=1.14.0
azure-monitor-opentelemetry-exporter==1.0.0b33